pkg_config

Function get_variable

pub fn get_variable(package: &str, variable: &str) -> Result<String, Error>

Run pkg-config to get the value of a variable from a package using --variable.

The content of PKG_CONFIG_SYSROOT_DIR is not injected in paths that are returned by pkg-config --variable, which makes them unsuitable to use during cross-compilation unless specifically designed to be used at that time.