Struct pkg_config::Config

pub struct Config {
    // some fields omitted
}

Methods

impl Config

fn new() -> Config

Creates a new set of configuration options which are all initially set to "blank".

fn statik(&mut self, statik: bool) -> &mut Config

Indicate whether the --static flag should be passed.

This will override the inference from environment variables described in the crate documentation.

fn atleast_version(&mut self, vers: &str) -> &mut Config

Indicate that the library must be at least version vers.

fn find(&self, name: &str) -> Result<Library, String>

Run pkg-config to find the library name.

This will use all configuration previously set to specify how pkg-config is run.