Struct pkg_config::Config

pub struct Config { /* fields omitted */ }

Methods

impl Config

pub fn new() -> Config

Creates a new set of configuration options which are all initially set to "blank".

pub fn statik(&mut self, statik: bool) -> &mut Config

Indicate whether the --static flag should be passed.

This will override the inference from environment variables described in the crate documentation.

pub fn atleast_version(&mut self, vers: &str) -> &mut Config

Indicate that the library must be at least version vers.

pub fn arg<S: AsRef<OsStr>>(&mut self, arg: S) -> &mut Config

Add an argument to pass to pkg-config.

It's placed after all of the arguments generated by this library.

pub fn cargo_metadata(&mut self, cargo_metadata: bool) -> &mut Config

Define whether metadata should be emitted for cargo allowing it to automatically link the binary. Defaults to true.

pub fn env_metadata(&mut self, env_metadata: bool) -> &mut Config

Define whether metadata should be emitted for cargo allowing to automatically rebuild when environment variables change. Defaults to false.

pub fn print_system_libs(&mut self, print: bool) -> &mut Config

Enable or disable the PKG_CONFIG_ALLOW_SYSTEM_LIBS environment variable.

This env var is enabled by default.

pub fn probe(&self, name: &str) -> Result<Library, Error>

Run pkg-config to find the library name.

This will use all configuration previously set to specify how pkg-config is run.